EGL: Fix a few errors in EGL_KHR_debug.
Fix the message types in the _eglReportError and _eglReportCritical macros.

In eglDebugMessageControlKHR, check if the attribute list is valid even if the
callback is NULL.

In eglDebugMessageControlKHR, set the callback pointer whether or not the
attribute list is NULL.

In eglLabelObjectKHR, if the vendor library returns an error for an object type
other than EGL_OBJECT_THREAD_KHR, then return that error to the caller.

Fix the error handling and return value in eglBindAPI if the API isn't one of
GL or GLES.

Add a missing error callback in eglMakeCurrent.