OpenGL: Statically export a few more APIs
ARB_multitexture and ARB_imaging have arguably been part of the OpenGL
ABI on Linux. Both are exported statically from Mesa's libGL, and
xserver's GLX assumes this. xserver would like to link against libOpenGL
instead of libGL so as not to drag client-side code into the server, but
the simple attempt to do this fails because libOpenGL does not expose
these APIs.

Signed-off-by: Adam Jackson <ajax@redhat.com>