GLX: Add GLX_SGIX_fbconfig functions
Some older apps, e.g. games using version 3.0 or older of the glfw library
try to use the old GLX_SGIX_fbconfig extension functions if present instead
of the regular glXGetFBConfigs, glXGetFBConfigAttrib and glXCreateNewContext
functions.

This commit adds implementations of these to libglvnd's libGLX fixing e.g.
the "The Binding if Isaac: Rebirth" and "Crypt of the NecroDancer" games from
Steam failing with a "GLX: Failed to retrieve visual for GLXFBConfig" error.

Note this commit does not add implementations for
glXCreateGLXPixmapWithConfigSGIX and glXGetFBConfigFromVisualSGIX as these
cannot be simply mapped to non extension functions and in practice these
do not seem to be used.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>