GLdispatch: Remove the ".syntax divided" directive for ARMv7.
The ".syntax divided" directive causes build errors some compilers.

The error is because the compiler apparently generates unified syntax by
default and expects inline assembly to use unified syntax, and so just
generates a single ".syntax unified" directive at the top.  The ".syntax
divided" directive then causes problems, because the compiler still generates
unified syntax for the rest of the C code in the file.

As far as I can tell, newer versions of gcc will expect inline assembly to use
divided syntax, and they generate the necessary .syntax directives to cope with
switching to unified syntax in inline assembly.