EGL: alias "drm" to gbm when using EGL_PLATFORM
Since the inseption of EGL/GBM the unofficial way to force
the platform was via EGL_PLATFORM="drm".

Strictly speaking one could remove the gbm one, but it should have no
side effects. Plus it's more consistent with other platforms.

In either case, one should not be using eglGetDisplay but the Platform
version of the API.

Signed-off-by: Emil Velikov <emil.velikov@gmail.com>