GLdispatch: Clean up the assembly dispatch code.
Made the names for macros and common variables more consistent.

Added a GLDISPATCH_PAGE_SIZE macro to specify the page size that it uses to
align the dispatch stubs. This value can be overridden at configure time, so a
later change could allow the configure script to figure out the page size on
architectures where it could vary.

This is all formatting changes, so there should be no functional difference.