Explain abbreviations and add links
This PR expands "GL" and other abbreviations, and adds links to Wikipedia so the reader may learn more about the history of the libraries. To a newcomer like me, there are a lot of abbreviations in the "Introduction" section of the document that are left unexplained, and I simply wanted to know what "GL" stood for but couldn't easily find this information. Since your library is one of the top hits for searching for "libgl," I thought it would be helpful to add this information here.