EGL: detect platform vendor Context invalidation method
Some vendors have implemented their drivers to the EGL 1.5 spec
such that Context pointer is invalid immediately following the
call to eglDestroyContext.  This means that providing the
application with the cached context state is invalid and breaks
the userspace application.

If the vendor returns true for this implementation rather than
using the cached values libglvnd should call back into the driver
to return its state.

Signed-off-by: Jon Nettleton <jon@solid-run.com>