EGL: internalize eglContextDestroy for state tracking
As noted before some vendors have followed the EGL 1.5 spec
such that as soon as eglContextDestroy is called on a context
it can no longer be passed for use by any other functions.
Consequently the same is true for Surfaces bound to the context.

In order to maintain that libglvnd is responsible for the state
of EGL, it is then necessary for libglvnd to take into account
this ambiguity in the EGL 1.5 spec.  Therefore if a vendor integration
specifies that invalidateContextOnDestroy is true, libglvnd needs
to take this into account and query the driver as to the current state
for the EGL thread.

This is implemented in such a way that existing platforms should not be
effected.

Signed-off-by: Jon Nettleton <jon@solid-run.com>