meson: Don't use x11 at all if the x11 feature is disabled.
Currently, building with -Dx11=disabled or -Dx11=auto gives identical
results.

In both cases, it only makes the x11 dependency optional: It'll still
look for libx11, and if libx11 is available, then it'll still build with
X11 support enabled.

This changes the meson build so that if you pass -Dx11=disabled, then it
will use a dummy dependency for x11, which will cause it to build as if
libx11 was not available.