EGL: Fix error reporting for EGL_PLATFORM_DEVICE_EXT
In eglGetPlatformDisplay, if the platform is EGL_PLATFORM_DEVICE_EXT and
the native display is a valid EGLDeviceEXT handle, then as an
optimization, libglvnd will dispatch to whichever vendor owns that
device, rather than iterating through all of them.

If the vendor's eglGetPlatformDisplay implementation returns
EGL_NO_DISPLAY, then libglvnd would return that to the caller, but it
does not correctly pass through the vendor's last EGL error code.

Add a call to __eglSetLastVendor in that case, so that the next call to
eglGetError() will correctly go to that vendor.