loongarch64: specify -mno-relax when supported
The LoongArch64 TSD assembly code relies on accurate .balign, which will
be broken by linker relaxation.

Specify -mno-relax when this option is supported. When this option is
not supported, assume linker relaxation is not available and nothing is
needed.

A new autoconf macro is introduced from autoconf-archive for easily
handling testing compiler options.

Signed-off-by: Icenowy Zheng <uwu@icenowy.me>